+++ .github/workflows/build.yml
@@ -12,6 +12,6 @@ jobs:
       - run: |
            docker build -f py.Dockerfile \
              --build-arg PYTHON_VERSION=${{ matrix.python-version }} \
-             --tag gym-minigrid-docker .      
+             --tag minigrid-docker .      
       - name: Run tests
-        run: docker run gym-minigrid-docker pytest
+        run: docker run minigrid-docker pytest
