--- gym_minigrid/envs/__init__.py
@@ -1,25 +1,0 @@
-from gym_minigrid.envs.blockedunlockpickup import BlockedUnlockPickupEnv
-from gym_minigrid.envs.crossing import CrossingEnv
-from gym_minigrid.envs.distshift import DistShiftEnv
-from gym_minigrid.envs.doorkey import DoorKeyEnv
-from gym_minigrid.envs.dynamicobstacles import DynamicObstaclesEnv
-from gym_minigrid.envs.empty import EmptyEnv
-from gym_minigrid.envs.fetch import FetchEnv
-from gym_minigrid.envs.fourrooms import FourRoomsEnv
-from gym_minigrid.envs.gotodoor import GoToDoorEnv
-from gym_minigrid.envs.gotoobject import GoToObjectEnv
-from gym_minigrid.envs.keycorridor import KeyCorridorEnv
-from gym_minigrid.envs.lavagap import LavaGapEnv
-from gym_minigrid.envs.lockedroom import LockedRoom, LockedRoomEnv
-from gym_minigrid.envs.memory import MemoryEnv
-from gym_minigrid.envs.multiroom import MultiRoom, MultiRoomEnv
-from gym_minigrid.envs.obstructedmaze import (
-    ObstructedMaze_1Dlhb,
-    ObstructedMaze_Full,
-    ObstructedMazeEnv,
-)
-from gym_minigrid.envs.playground import PlaygroundEnv
-from gym_minigrid.envs.putnear import PutNearEnv
-from gym_minigrid.envs.redbluedoors import RedBlueDoorEnv
-from gym_minigrid.envs.unlock import UnlockEnv
-from gym_minigrid.envs.unlockpickup import UnlockPickupEnv
