+++ minigrid/envs/blockedunlockpickup.py
@@ -1,40 +1,45 @@
-from gym_minigrid.minigrid import COLOR_NAMES, MissionSpace
-from gym_minigrid.roomgrid import RoomGrid
+from minigrid.minigrid import COLOR_NAMES, Ball, MissionSpace
+from minigrid.roomgrid import RoomGrid
 
 
-class UnlockPickupEnv(RoomGrid):
+class BlockedUnlockPickupEnv(RoomGrid):
 
     """
     ### Description
 
     The agent has to pick up a box which is placed in another room, behind a
-    locked door. This environment can be solved without relying on language.
+    locked door. The door is also blocked by a ball which the agent has to move
+    before it can unlock the door. Hence, the agent has to learn to move the
+    ball, pick up the key, open the door and pick up the object in the other
+    room. This environment can be solved without relying on language.
 
     ### Mission Space
 
-    "pick up the {color} box"
+    "pick up the {color} {type}"
 
     {color} is the color of the box. Can be "red", "green", "blue", "purple",
     "yellow" or "grey".
 
+    {type} is the type of the object. Can be "box" or "key".
+
     ### Action Space
 
-    | Num | Name         | Action                    |
-    |-----|--------------|---------------------------|
-    | 0   | left         | Turn left                 |
-    | 1   | right        | Turn right                |
-    | 2   | forward      | Move forward              |
-    | 3   | pickup       | Pick up an object         |
-    | 4   | drop         | Unused                    |
-    | 5   | toggle       | Toggle/activate an object |
-    | 6   | done         | Unused                    |
+    | Num | Name         | Action            |
+    |-----|--------------|-------------------|
+    | 0   | left         | Turn left         |
+    | 1   | right        | Turn right        |
+    | 2   | forward      | Move forward      |
+    | 3   | pickup       | Pick up an object |
+    | 4   | drop         | Unused            |
+    | 5   | toggle       | Unused            |
+    | 6   | done         | Unused            |
 
     ### Observation Encoding
 
     - Each tile is encoded as a 3 dimensional tuple:
         `(OBJECT_IDX, COLOR_IDX, STATE)`
     - `OBJECT_TO_IDX` and `COLOR_TO_IDX` mapping can be found in
-        [gym_minigrid/minigrid.py](gym_minigrid/minigrid.py)
+        [minigrid/minigrid.py](minigrid/minigrid.py)
     - `STATE` refers to the door state with 0=open, 1=closed and 2=locked
 
     ### Rewards
@@ -50,22 +55,22 @@ class UnlockPickupEnv(RoomGrid):
 
     ### Registered Configurations
 
-    - `MiniGrid-Unlock-v0`
+    - `MiniGrid-BlockedUnlockPickup-v0`
 
     """
 
     def __init__(self, **kwargs):
         room_size = 6
         mission_space = MissionSpace(
-            mission_func=lambda color: f"pick up the {color} box",
-            ordered_placeholders=[COLOR_NAMES],
+            mission_func=lambda color, type: f"pick up the {color} {type}",
+            ordered_placeholders=[COLOR_NAMES, ["box", "key"]],
         )
         super().__init__(
             mission_space=mission_space,
             num_rows=1,
             num_cols=2,
             room_size=room_size,
-            max_steps=8 * room_size**2,
+            max_steps=16 * room_size**2,
             **kwargs,
         )
 
@@ -75,7 +80,10 @@ class UnlockPickupEnv(RoomGrid):
         # Add a box to the room on the right
         obj, _ = self.add_object(1, 0, kind="box")
         # Make sure the two rooms are directly connected by a locked door
-        door, _ = self.add_door(0, 0, 0, locked=True)
+        door, pos = self.add_door(0, 0, 0, locked=True)
+        # Block the door with a ball
+        color = self._rand_color()
+        self.grid.set(pos[0] - 1, pos[1], Ball(color))
         # Add a key to unlock the door
         self.add_object(0, 0, "key", door.color)
