+++ minigrid/envs/dynamicobstacles.py
@@ -2,7 +2,7 @@ from operator import add
 
 from gym.spaces import Discrete
 
-from gym_minigrid.minigrid import Ball, Goal, Grid, MiniGridEnv, MissionSpace
+from minigrid.minigrid import Ball, Goal, Grid, MiniGridEnv, MissionSpace
 
 
 class DynamicObstaclesEnv(MiniGridEnv):
@@ -37,7 +37,7 @@ class DynamicObstaclesEnv(MiniGridEnv):
     - Each tile is encoded as a 3 dimensional tuple:
         `(OBJECT_IDX, COLOR_IDX, STATE)`
     - `OBJECT_TO_IDX` and `COLOR_TO_IDX` mapping can be found in
-        [gym_minigrid/minigrid.py](gym_minigrid/minigrid.py)
+        [minigrid/minigrid.py](minigrid/minigrid.py)
     - `STATE` refers to the door state with 0=open, 1=closed and 2=locked
 
     ### Rewards
