+++ minigrid/envs/keycorridor.py
@@ -1,5 +1,5 @@
-from gym_minigrid.minigrid import COLOR_NAMES, MissionSpace
-from gym_minigrid.roomgrid import RoomGrid
+from minigrid.minigrid import COLOR_NAMES, MissionSpace
+from minigrid.roomgrid import RoomGrid
 
 
 class KeyCorridorEnv(RoomGrid):
@@ -41,7 +41,7 @@ class KeyCorridorEnv(RoomGrid):
     - Each tile is encoded as a 3 dimensional tuple:
         `(OBJECT_IDX, COLOR_IDX, STATE)`
     - `OBJECT_TO_IDX` and `COLOR_TO_IDX` mapping can be found in
-        [gym_minigrid/minigrid.py](gym_minigrid/minigrid.py)
+        [minigrid/minigrid.py](minigrid/minigrid.py)
     - `STATE` refers to the door state with 0=open, 1=closed and 2=locked
 
     ### Rewards
