+++ minigrid/envs/lockedroom.py
@@ -1,4 +1,4 @@
-from gym_minigrid.minigrid import (
+from minigrid.minigrid import (
     COLOR_NAMES,
     Door,
     Goal,
@@ -59,7 +59,7 @@ class LockedRoomEnv(MiniGridEnv):
     - Each tile is encoded as a 3 dimensional tuple:
         `(OBJECT_IDX, COLOR_IDX, STATE)`
     - `OBJECT_TO_IDX` and `COLOR_TO_IDX` mapping can be found in
-        [gym_minigrid/minigrid.py](gym_minigrid/minigrid.py)
+        [minigrid/minigrid.py](minigrid/minigrid.py)
     - `STATE` refers to the door state with 0=open, 1=closed and 2=locked
 
     ### Rewards
