+++ minigrid/envs/unlockpickup.py
@@ -1,18 +1,21 @@
-from gym_minigrid.minigrid import MissionSpace
-from gym_minigrid.roomgrid import RoomGrid
+from minigrid.minigrid import COLOR_NAMES, MissionSpace
+from minigrid.roomgrid import RoomGrid
 
 
-class UnlockEnv(RoomGrid):
+class UnlockPickupEnv(RoomGrid):
 
     """
     ### Description
 
-    The agent has to open a locked door. This environment can be solved without
-    relying on language.
+    The agent has to pick up a box which is placed in another room, behind a
+    locked door. This environment can be solved without relying on language.
 
     ### Mission Space
 
-    "open the door"
+    "pick up the {color} box"
+
+    {color} is the color of the box. Can be "red", "green", "blue", "purple",
+    "yellow" or "grey".
 
     ### Action Space
 
@@ -21,7 +24,7 @@ class UnlockEnv(RoomGrid):
     | 0   | left         | Turn left                 |
     | 1   | right        | Turn right                |
     | 2   | forward      | Move forward              |
-    | 3   | pickup       | Unused                    |
+    | 3   | pickup       | Pick up an object         |
     | 4   | drop         | Unused                    |
     | 5   | toggle       | Toggle/activate an object |
     | 6   | done         | Unused                    |
@@ -31,7 +34,7 @@ class UnlockEnv(RoomGrid):
     - Each tile is encoded as a 3 dimensional tuple:
         `(OBJECT_IDX, COLOR_IDX, STATE)`
     - `OBJECT_TO_IDX` and `COLOR_TO_IDX` mapping can be found in
-        [gym_minigrid/minigrid.py](gym_minigrid/minigrid.py)
+        [minigrid/minigrid.py](minigrid/minigrid.py)
     - `STATE` refers to the door state with 0=open, 1=closed and 2=locked
 
     ### Rewards
@@ -42,7 +45,7 @@ class UnlockEnv(RoomGrid):
 
     The episode ends if any one of the following conditions is met:
 
-    1. The agent opens the door.
+    1. The agent picks up the correct box.
     2. Timeout (see `max_steps`).
 
     ### Registered Configurations
@@ -53,19 +56,24 @@ class UnlockEnv(RoomGrid):
 
     def __init__(self, **kwargs):
         room_size = 6
-        mission_space = MissionSpace(mission_func=lambda: "open the door")
+        mission_space = MissionSpace(
+            mission_func=lambda color: f"pick up the {color} box",
+            ordered_placeholders=[COLOR_NAMES],
+        )
         super().__init__(
             mission_space=mission_space,
             num_rows=1,
             num_cols=2,
             room_size=room_size,
             max_steps=8 * room_size**2,
-            **kwargs
+            **kwargs,
         )
 
     def _gen_grid(self, width, height):
         super()._gen_grid(width, height)
 
+        # Add a box to the room on the right
+        obj, _ = self.add_object(1, 0, kind="box")
         # Make sure the two rooms are directly connected by a locked door
         door, _ = self.add_door(0, 0, 0, locked=True)
         # Add a key to unlock the door
@@ -73,14 +81,14 @@ class UnlockEnv(RoomGrid):
 
         self.place_agent(0, 0)
 
-        self.door = door
-        self.mission = "open the door"
+        self.obj = obj
+        self.mission = f"pick up the {obj.color} {obj.type}"
 
     def step(self, action):
         obs, reward, terminated, truncated, info = super().step(action)
 
-        if action == self.actions.toggle:
-            if self.door.is_open:
+        if action == self.actions.pickup:
+            if self.carrying and self.carrying == self.obj:
                 reward = self._reward()
                 terminated = True
