+++ tests/utils.py
@@ -5,7 +5,7 @@ import numpy as np
 all_testing_env_specs = [
     env_spec
     for env_spec in gym.envs.registry.values()
-    if env_spec.entry_point.startswith("gym_minigrid.envs")
+    if env_spec.entry_point.startswith("minigrid.envs")
 ]
